+++ Assets/Scripts/Data/CombatPhaseData.cs
@@ -7,7 +7,7 @@ namespace BeatHero.Data
     [Serializable]
     public class CombatPhaseData
     {
-        public int bpm;
+        public float bpm;
         public AudioClip bgm;
         public List<PatternData> patterns;
     }
